move serial wrappers in vmm/device module
We will add wrappers for the RTC as well, so it makes sense to have them
in a separate module, so we don't have an ever growing flattened vmm
module.

As part of this change, also updated vm-superio to version 0.2 and
adapted the existing wrappers. This is needed to be able to consume RTC
(as it is only available starting with 0.2).
